requirements.txt
playwright==1.40.0
openai==1.3.0
pyyaml==6.0.1
python-dotenv==1.0.0
pillow==10.1.0
requests==2.31.0 